move walker code to own package
