+++ pkg/util/trivy/walker/testdata/fs/bar
